[CT-23] refactor: 파일분리
